+++ exiftags/exifutil.c
@@ -89,7 +89,8 @@ offsanity(struct exifprop *prop, u_int16_t size, struct ifd *dir)
 	const char *name;
 
 	/* XXX Hrm.  Should be OK with 64-bit addresses. */
-	tifflen = dir->md.etiff - dir->md.btiff;
+	// assuming no TIFF file is so big that the difference in offsets will be greater than UINT32_MAX
+	tifflen = (u_int32_t)(dir->md.etiff - dir->md.btiff);
 	if (prop->name)
 		name = prop->name;
 	else
@@ -315,7 +316,7 @@ childprop(struct exifprop *parent)
  * Allocate a buffer for a property's display string.
  */
 void
-exifstralloc(char **str, int len)
+exifstralloc(char **str, size_t len)
 {
 
 	if (*str) {
@@ -384,7 +385,7 @@ readifd(u_int32_t offset, struct ifd **dir, struct exiftag *tagset,
 	unsigned char *b;
 	struct ifdoff *ifdoffs, *lastoff;
 
-	tifflen = md->etiff - md->btiff;
+	tifflen = (u_int32_t)(md->etiff - md->btiff);
 	b = md->btiff;
 	ifdoffs = (struct ifdoff *)(md->ifdoffs);
 	lastoff = NULL;
